Kernel se comunica bien con memoria pero sin usar las funciones de protocol

+++ memoria/src/memoria_utils.h
@@ -51,7 +51,7 @@ void apagar_memoria();
 void escuchar_kernel();
 void recibir_kernel();
 
-uint32_t crear_tablas(uint16_t pid, uint32_t tamanio);
+uint32_t crear_tablas(uint16_t pid, uint16_t tamanio);
 void inicializar_tabla_1er_nivel(fila_1er_nivel* tabla_1er_nivel);
 void inicializar_tabla_2do_nivel(fila_2do_nivel* tabla_2do_nivel);
 
@@ -84,7 +84,7 @@ void* obtener_marco(uint32_t nro_marco);
 uint32_t marcos_en_memoria();
 void escribir_marco_en_memoria(uint32_t nro_marco, void* marco);
 int buscar_marco_libre();
-uint32_t calcular_cant_marcos(uint32_t tamanio);
+uint32_t calcular_cant_marcos(uint16_t tamanio);
 int marcos_actuales(int entrada_1er_nivel, int entrada_2do_nivel);
 void printear_bitmap();
